+++ tests/test_app/TestApp/config/routes.php
@@ -2,6 +2,6 @@
 use Cake\Routing\Router;
 
 Router::scope('/', function ($routes) {
-    $routes->loadPlugin(THUMBER);
+    $routes->loadPlugin('Thumber');
     $routes->loadPlugin('MeCmsInstagram');
 });