+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 1.x branch
 ## 1.9 branch
 ### 1.9.0
-* removed `ME_CMS_INSTAGRAM` constants. It no longer uses also `ME_CMS` constant.
+* removed `ME_CMS_INSTAGRAM` constants. It no longer uses also `ME_CMS` and
+    `THUMBER` constants;
+* updated for CakePHP 3.7.
 
 ## 1.8 branch
 ### 1.8.1